Add tests for get_random_character()
The tests check that:

- the function returns a single lowercase character
- all 26 lowercase characters are possible
- the distribution of characters is fair and uniform (p<0.05)

To do that last test, I had to implement `chi_squared_uniform()`. This
required two further tests. One test just feeds the function a `Vec` of
ones (which is perfectly uniform and should return `0.0`), and the other
feeds it 50 random (pre-chosen and static) `u8` values and checks the
result against a pre-calculated statistic value.